add "-undefined dynamic_lookup" option to plugins
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@8779 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/fileloader/scribus12format/scribus12format.h
@@ -57,8 +57,8 @@ class PLUGIN_API Scribus12Format : public LoadSavePlugin
 //		bool newReplacement;
 //		QMap<QString,QString> ReplacedFonts;
 //		QMap<QString,QString> DoFonts;
-		QMap<long,int> itemRemap;
-		QMap<int,long> itemNext;
+		QMap<long long,int> itemRemap;
+		QMap<int,long long> itemNext;
 		QMap<uint,QString> DoVorl;
 		uint VorlC;
